+++ src/editor_window.rs
@@ -22,7 +22,7 @@ pub(crate) fn yoleck_editor_window(
                     world.resource_scope(|world, mut edit_specific: Mut<EditSpecificResources>| {
                         edit_specific.inject_to_world(world);
                         for section in yoleck_editor_sections.0.iter_mut() {
-                            section.0.invoke(world, ui);
+                            section.0.invoke(world, ui).unwrap();
                         }
                         edit_specific.take_from_world(world);
                     });
@@ -43,23 +43,24 @@ enum YoleckEditorSectionInner {
                 + Sync
                 + FnOnce(
                     &mut World,
-                )
-                    -> Box<dyn 'static + Send + Sync + FnMut(&mut World, &mut egui::Ui)>,
+                ) -> Box<
+                    dyn 'static + Send + Sync + FnMut(&mut World, &mut egui::Ui) -> Result,
+                >,
         >,
     ),
     Middle,
-    Initialized(Box<dyn 'static + Send + Sync + FnMut(&mut World, &mut egui::Ui)>),
+    Initialized(Box<dyn 'static + Send + Sync + FnMut(&mut World, &mut egui::Ui) -> Result>),
 }
 
 impl YoleckEditorSectionInner {
-    fn invoke(&mut self, world: &mut World, ui: &mut egui::Ui) {
+    fn invoke(&mut self, world: &mut World, ui: &mut egui::Ui) -> Result {
         match self {
             Self::Uninitialized(_) => {
                 if let Self::Uninitialized(system_constructor) =
                     core::mem::replace(self, Self::Middle)
                 {
                     let mut system = system_constructor(world);
-                    system(world, ui);
+                    system(world, ui)?;
                     *self = Self::Initialized(system);
                 } else {
                     panic!("It was just Uninitialized...");
@@ -67,9 +68,10 @@ impl YoleckEditorSectionInner {
             }
             Self::Middle => panic!("Cannot start in the middle state when being invoked"),
             Self::Initialized(system) => {
-                system(world, ui);
+                system(world, ui)?;
             }
         }
+        Ok(())
     }
 }
 
@@ -79,7 +81,7 @@ pub struct YoleckEditorSection(YoleckEditorSectionInner);
 impl<C, S> From<C> for YoleckEditorSection
 where
     C: 'static + Send + Sync + FnOnce(&mut World) -> S,
-    S: 'static + Send + Sync + FnMut(&mut World, &mut egui::Ui),
+    S: 'static + Send + Sync + FnMut(&mut World, &mut egui::Ui) -> Result,
 {
     fn from(system_constructor: C) -> Self {
         Self(YoleckEditorSectionInner::Uninitialized(Box::new(